refactor(apps/app): 로그인 로직 훅으로 분리
